backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
ollama==0.4.3
python-dotenv==1.0.1
numpy
stripe
anthropic
httpx
